requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.8.0; python_version >= '3.9'
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.1.31; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
db-dtypes==1.4.2; python_version >= '3.7'
exceptiongroup==1.2.2; python_version >= '3.7'
fastapi==0.115.11; python_version >= '3.8'
google-api-core[grpc]==2.24.2; python_version >= '3.7'
google-auth==2.38.0; python_version >= '3.7'
google-cloud-bigquery==3.30.0; python_version >= '3.7'
google-cloud-core==2.4.3; python_version >= '3.7'
google-crc32c==1.6.0; python_version >= '3.9'
google-resumable-media==2.7.2; python_version >= '3.7'
googleapis-common-protos==1.69.1; python_version >= '3.7'
grpcio==1.71.0; python_version >= '3.9'
grpcio-status==1.71.0; python_version >= '3.9'
h11==0.14.0; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
numpy==2.0.2; python_version >= '3.9'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
proto-plus==1.26.1; python_version >= '3.7'
protobuf==5.29.3; python_version >= '3.8'
pyarrow==19.0.1; python_version >= '3.9'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.1; python_version >= '3.8'
pydantic==2.10.6; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-multipart==0.0.20; python_version >= '3.8'
pytz==2025.1
requests==2.32.3; python_version >= '3.8'
rsa==4.9; python_version >= '3.6' and python_version < '4'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.46.1; python_version >= '3.9'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2025.1; python_version >= '2'
urllib3==2.3.0; python_version >= '3.9'
uvicorn==0.34.0; python_version >= '3.9'
